🎨 Improve code of page & worker scripts
